+++ cypress/locators/row/index.js
@@ -1,5 +1,0 @@
-import { ROW, COLUMN } from "./locators";
-
-// component preview locators
-export const row = () => cy.get(ROW);
-export const column = () => row().find(COLUMN);