+++ src/components/modal/modal.component.js
@@ -175,6 +175,7 @@ const Modal = ({
                 value={{
                   isAnimationComplete,
                   triggerRefocusFlag,
+                  isInModal: true,
                 }}
               >
                 {content}